+++ wies/core/services/users.py
@@ -65,16 +65,21 @@ def create_user(creator: User, first_name, last_name, email, labels=None, groups
     if groups:
         user.groups.set(groups)
 
-    creator_user_email = creator.email if creator is not None else ""
     context = {
-        "created_id": user.id,
         "email": email,
         "first_name": first_name,
         "last_name": last_name,
         "label_names": label_names,
         "group_names": [group.name for group in groups],
     }
-    create_event(creator_user_email, "User.create", context=context)
+    create_event(
+        object_type="User",
+        action="create",
+        source="user",
+        object_id=user.id,
+        user=creator,
+        context=context,
+    )
 
     return user
 
@@ -119,16 +124,21 @@ def update_user(updater, user, first_name, last_name, email, labels=None, groups
     if groups:
         user.groups.set(groups)
 
-    updater_user_email = updater.email if updater is not None else ""
     context = {
-        "updated_id": user.id,
         "first_name": first_name,
         "last_name": last_name,
         "email": email,
         "label_names": label_names,
         "group_names": [group.name for group in groups],
     }
-    create_event(updater_user_email, "User.update", context=context)
+    create_event(
+        object_type="User",
+        action="update",
+        source="user",
+        object_id=user.id,
+        user=updater,
+        context=context,
+    )
     return user